+++ orchestrator/src/core/client/queue/sqs.rs
@@ -446,7 +446,7 @@ impl QueueClient for SQS {
         Ok(())
     }
 
-    async fn get_queue_depth(&self, queue: QueueType) -> Result<i32, QueueError> {
+    async fn get_queue_depth(&self, queue: QueueType) -> Result<usize, QueueError> {
         let queue_name = self.get_queue_name(&queue)?;
         let queue_url = self.inner.get_queue_url_from_client(queue_name.as_str()).await?;
 
@@ -462,7 +462,7 @@ impl QueueClient for SQS {
         let count = attributes
             .attributes()
             .and_then(|attrs| attrs.get(&QueueAttributeName::ApproximateNumberOfMessages))
-            .and_then(|value| value.parse::<i32>().ok())
+            .and_then(|value| value.parse::<usize>().ok())
             .unwrap_or(0);
 
         Ok(count)
